disintegration

+++ src/lib/potree-loader/materials/shaders/pointcloud.frag
@@ -1,5 +1,5 @@
 precision highp float;
-precision mediump int;
+precision highp int;
 
 uniform mat4 viewMatrix;
 uniform vec3 cameraPosition;
